修改页面bug 修改头部导航栏

+++ app/Http/Controllers/Admin/ArtController.php
@@ -101,8 +101,11 @@ class ArtController extends Controller
     public function del($art_id)
     {
         $filename = DB::table('article')->where('art_id',$art_id)->select('art_image')->first();
-        unlink($filename->art_image);
+        if ($filename->art_image){
+            unlink($filename->art_image);
+        }
         $re = DB::table('article')->where('art_id',$art_id)->delete();
+
         if ($re){
             $data = [
                 'state' => '200',
